// src/app/interfaces/timeline-item.ts
export interface TimelineItem {
    label?: string;
    icon?: string;
    command?: (event?: any) => void;
    url?: string;
    routerLink?: any;
    expanded?: boolean;
    disabled?: boolean;
    visible?: boolean;
    active?: boolean;
    style?: any;
    color?: string;
    styleClass?: string;
    title?: string;
    id?: string;
  }